+++ injector/src/index.js
@@ -9,8 +9,6 @@ import CSP from "./modules/csp";
 if (!process.argv.includes("--vanilla")) {
     process.env.NODE_OPTIONS = "--no-force-async-hooks-checks";
     app.commandLine.appendSwitch("no-force-async-hooks-checks");
-    process.electronBinding("command_line").appendSwitch("no-force-async-hooks-checks");
-
 
     // Patch and replace the built-in BrowserWindow
     BrowserWindow.patchBrowserWindow();